# Bug 1177619 - Score div nodes which have br nodes. r=Gijs