--- a/test/test-pages/blogger/expected-metadata.json
+++ b/test/test-pages/blogger/expected-metadata.json
@@ -1,0 +1,6 @@
+{
+  "title": "Open Verilog flow for Silego GreenPak4 programmable logic devices",
+  "byline": null,
+  "excerpt": "I've written a couple of posts in the past few months but they were all for",
+  "readerable": true
+}