--- a/test/test-pages/replace-brs/expected-metadata.json
+++ b/test/test-pages/replace-brs/expected-metadata.json
@@ -2,5 +2,5 @@
   "title": "Replace brs test",
   "byline": null,
   "excerpt": "Lorem ipsum",
-  "readerable": false
+  "readerable": true
 }