The solution here is score `<br>` tags because blogger use `<br>` tags to typeset their articles not `<p>` tags.

Their DOM structure looks like

```
<div>
  Dare to predict and suggest and improve your prediction based on time. Always try to add some more value to the user.<br />
  Its a continuous process, reiterate the above point &nbsp;time to time.<br />
  Let me know your thoughts how to get a wow ....<br />
  <br />
  Dare to predict and suggest and improve your prediction based on time. Always try to add some more value to the user.<br />
  Its a continuous process, reiterate the above point &nbsp;time to time.<br />
  Let me know your thoughts how to get a wow ....<br />
</div>
```

I also added tests. And the test web page is from http://www.myexperimentswithjava.com/2016/06/magic-with-computers.html.