@evanxd 2016-10-11
The CI error is about JSDOMParser error. We can fix it by clean up the web page. Before that, let's make sure the approach of the patch is good.