@evanxd 2016-10-20
Hi @gijsk ,

I've fixed the test failures and done the benchmark experiments. 

The result(with jsdom)[1] is 38.46% slower after patched the patch[3]. Additionally, the result with JSDOMParser[2] is 24.38% slower. You can check the detail on the reports[1][2].

I Run `npm run perf-reference` for 10 times for each experiment.  And the slower rate = `(Patched Result - Original Result)/Original Result`.

What do you think of the benchmark result? Could it be acceptable?

[1]: https://docs.google.com/spreadsheets/d/1s8CXR8QCdSZuAyrVvuYM8Zh-4aq6H9ZRP5g6wpKHIi0
[2]: https://docs.google.com/spreadsheets/d/1I4S0yQZ4-I9p3YI3NIP4jiuSasL3qDtp4QGEcKuqt7Y
[3]: https://github.com/mozilla/readability/pull/310/commits/b5602a4b715efb9211200af3b7660d0929005ed6