@evanxd 2016-10-21
Hi @gijsk ,

I've updated patch[1] for your comments. And the benchmark(with `jsdom`) result[2] is `36.87%` slower. You can check the detail of benchmark experiment on the report[2]. 

I Run `npm run perf-reference` for 10 times for each experiment. And the slower rate = `(Patched Result - Original Result)/Original Result`.

What do you think of the benchmark result of `isProbablyReaderable ` method(36.87% slower)? Could it be acceptable?

[1]: https://github.com/mozilla/readability/pull/310/commits/61af7898401869d49a0ffbb1e0c769b6b8893bd6
[2]: https://docs.google.com/spreadsheets/d/140SfoJ3YTkExhrsBgFSuL8SWTinPNY_kVl7TS7EfQ6U