@gijsk 2016-10-21
37% is a lot. Especially because this is a decrease in ops/s, and so really what it means is that the operation itself is now 60% slower (ie the same operation is now 60% slower, so if it used to take e.g. 100ms to run it now takes 160, which means that on a sample time of 1000ms we run ~6.4 ops instead of 10, which looks like a 36% decrease in ops/s but is actually bigger in terms of absolute speed in ms.).

I don't know how `jsdom`'s querySelectorAll performance compares with that of the native DOMs we use. Do you think it would be possible to make the modification to Readability.js from the patch to the one in the tree, and then use the browser console to run some rough experiments on a few pages? I recognize this is more work than running the commandline benchmark, but it would be a more accurate estimate of what the perf impact will be in practice.