@gijsk 2016-10-26
I guess we might have to take the hit, then, yeah. :-(

It would be useful if we had a way of avoiding the hit. Perhaps we can file a followup bug to optimize this method? I bet that either running these checks sequentially or unifying the different selectors might help.