@gijsk 2016-10-26
Though to be fair, a 30% hit on a br-heavy page is a lot better than what the jsdom tests suggested.