@evanxd 2016-10-27
Sure, I've filed a followup bug[1] to optimize the method.

So what's our next step for the patch? Just start to write tests with a blogger page(maybe use the page[2] with many `<br>` tags)?

Looks like we're almost done here. :-)

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=1313301
[2]: http://siliconexposed.blogspot.tw/2016/05/open-verilog-flow-for-silego-greenpak4.html