@gijsk 2016-10-28
Yes, add a test, make sure tests pass, and update any styling nits (haven't looked at the patch again), then we should be done.