@evanxd 2016-10-28
Got it. Let me add the tests.

One more question, update what styling nits? It seems that we haven't discussed any (JS code) styling nits before for the patch.