@evanxd 2016-11-01
Weird, the eslint failures seems not be caused by the patch. Investigate...

```
/home/travis/build/mozilla/readability/Readability.js
  1846:14  error  Expected indentation of 6 spaces but found 13  indent
  1847:14  error  Expected indentation of 6 spaces but found 13  indent
  1848:14  error  Expected indentation of 6 spaces but found 13  indent
  1849:14  error  Expected indentation of 6 spaces but found 13  indent
  1850:14  error  Expected indentation of 6 spaces but found 13  indent
  1851:14  error  Expected indentation of 6 spaces but found 13  indent
  1852:14  error  Expected indentation of 6 spaces but found 13  indent
```
https://travis-ci.org/mozilla/readability/builds/172231876#L397-L404