@gijsk 2016-11-01
I'm pretty reluctant to fix the eslint version because IME this ends up meaning you get stuck on outdated versions if you're not paying attention (which we tend not to :-) ). So I'd just leave it as-is and fix the indentation while we're here.