@evanxd 2016-11-01
Sure, got it and agreed.

Added the trailing comma for your comment. Thanks for reviewing. :-)